Move MapDrawingInterface functions to library
